Consolidation of path variables into config file.
Using a quasi-global root variable in the config object that sets the
paths for all dependencies. The dependent objects just have to load the
config object, which is always at the root of the application. If it has
any dependencies on other files, it needs to find them via the config
object.